requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
gunicorn==20.1.0
huggingface-hub==0.8.1
idna==3.3
numpy==1.23.1
packaging==21.3
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
sqlparse==0.4.2
huggingface-hub==0.8.1
idna==3.3
joblib==1.1.0
numpy==1.23.1
packaging==21.3
pandas==1.4.3
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.0
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.10
